+++ examples/apriori.py
@@ -5,8 +5,8 @@ from datetime import timedelta
 
 from bytewax.connectors.files import FileInput
 from bytewax.dataflow import Dataflow
-from bytewax.execution import run_main, spawn_cluster
-from bytewax.outputs import ManualOutputConfig, StdOutputConfig, TestingOutputConfig
+from bytewax.execution import run_main
+from bytewax.outputs import TestingOutputConfig
 from bytewax.window import SystemClockConfig, TumblingWindowConfig