+++ examples/events_to_parquet.py
@@ -9,12 +9,12 @@ from pyarrow import parquet, Table
 
 from bytewax.dataflow import Dataflow
 from bytewax.execution import spawn_cluster
-from bytewax.inputs import CustomPartInput
+from bytewax.inputs import PartInput
 from bytewax.outputs import ManualOutputConfig
 from bytewax.window import SystemClockConfig, TumblingWindowConfig
 
 
-class FakeWebEventsInput(CustomPartInput):
+class FakeWebEventsInput(PartInput):
     def list_parts(self):
         return ["singleton"]