+++ src/inputs.rs
@@ -3,14 +3,13 @@
 //! For a user-centric version of input, read the `bytewax.inputs`
 //! Python module docstring. Read that first.
 //!
-//! [`CustomPartInput`] defines an input source. [`PartBundle`] is
-//! what is passed to the source operators defined in
+//! [`PartInput`] defines an input source. [`PartBundle`] is what is
+//! passed to the source operators defined in
 //! [`crate::execution::epoch`].
 //!
 //! Each [`PartIter`] is keyed by a [`StateKey`] so that the recovery
 //! system can round trip the state data back to
-//! [`CustomPartInput::build_parts`] and be provided to the correct
-//! builder.
+//! [`PartInput::build_parts`] and be provided to the correct builder.
 //!
 //! The one extra quirk here is that input is completely decoupled
 //! from epoch generation. See [`crate::execution`] for how Timely
@@ -29,21 +28,21 @@ use pyo3::exceptions::PyTypeError;
 use pyo3::prelude::*;
 use pyo3::types::PyIterator;
 
-/// Represents a `bytewax.inputs.CustomPartInput` from Python.
+/// Represents a `bytewax.inputs.PartInput` from Python.
 #[derive(Clone)]
-pub(crate) struct CustomPartInput(Py<PyAny>);
+pub(crate) struct PartInput(Py<PyAny>);
 
 /// Do some eager type checking.
-impl<'source> FromPyObject<'source> for CustomPartInput {
+impl<'source> FromPyObject<'source> for PartInput {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
         let abc = ob
             .py()
             .import("bytewax.inputs")?
-            .getattr("CustomPartInput")?
+            .getattr("PartInput")?
             .extract()?;
         if !ob.is_instance(abc)? {
             Err(PyTypeError::new_err(
-                "input must derive from `bytewax.inputs.CustomPartInput`",
+                "input must derive from `bytewax.inputs.PartInput`",
             ))
         } else {
             Ok(Self(ob.into()))
@@ -51,7 +50,7 @@ impl<'source> FromPyObject<'source> for CustomPartInput {
     }
 }
 
-impl IntoPy<Py<PyAny>> for CustomPartInput {
+impl IntoPy<Py<PyAny>> for PartInput {
     fn into_py(self, _py: Python<'_>) -> Py<PyAny> {
         self.0
     }
@@ -62,7 +61,7 @@ impl IntoPy<Py<PyAny>> for CustomPartInput {
 /// Not just on this worker.
 pub(crate) struct TotalPartCount(pub(crate) usize);
 
-impl CustomPartInput {
+impl PartInput {
     /// Build all partitions for this input for this worker.
     pub(crate) fn build_parts(
         &self,